libX11 mesa git update 22 Jul 2013
libX11           commit 24d3ee0d08f24e23c91d55702f010f73d7b908e5
mesa             commit 190312949e8ce2c1dc884d4db5d6a44511666641

+++ libX11/src/FontNames.c
@@ -66,7 +66,7 @@ int *actualCount)	/* RETURN */
 
     if (rep.nFonts) {
 	flist = Xmalloc (rep.nFonts * sizeof(char *));
-	if (rep.length < (LONG_MAX >> 2)) {
+	if (rep.length < (INT_MAX >> 2)) {
 	    rlen = rep.length << 2;
 	    ch = Xmalloc(rlen + 1);
 	    /* +1 to leave room for last null-terminator */
